+++ R/encounter.R
@@ -201,3 +201,24 @@ encounter.ctmm <- function(CTMM,include=NULL,exclude=NULL,debias=FALSE,...)
   CTMM <- ctmm(mu=mu,sigma=sigma,COV.mu=COV.mu,COV=COV,axes=axes,isotropic=isotropic,info=info)
   return(list(CTMM=CTMM,BIAS=BIAS,bias=bias))
 }
+
+
+# relative encounter rates
+rates <- function(object,debias=TRUE,level=0.95,normalize=TRUE,...)
+{
+  units <- FALSE
+
+  R <- overlap(object,debias=debias,level=level,method="Rate",...)
+
+  if(normalize)
+  {
+    M <- 1/diag(R[,,'est']) # TODO !!! REMOVE MEAN BIAS CORRECTION
+    M <- mean(M) # TODO !!! REPLACE THIS WITH META-MEAN
+    R <- R*M
+  }
+
+  # fix diagonals # self encounter rate
+  diag(R[,,1]) <- diag(R[,,2]) <- diag(R[,,3]) <- Inf
+
+  return(R)
+}